requirements.txt
feedparser
python-telegram-bot